✅ Add a shared mockWebSocket test utility
Lift the fake WebSocket inlined in the WebSocket observable spec into a shared
core test helper, so upcoming WebSocket tickets drive instrumentation through
one test double instead of growing their own copies.

The utility follows the XHR mock's shape: it swaps the `WebSocket` global and
registers its own cleanup, including resetting the observable singleton. The
observable spec's test cases and assertions are unchanged.
