+++ packages/browser-core/test/index.ts
@@ -17,6 +17,7 @@ export * from './emulate/mockEventBridge'
 export * from './emulate/mockFlushController'
 export * from './emulate/mockFetch'
 export * from './emulate/mockXhr'
+export * from './emulate/mockWebSocket'
 export * from './emulate/mockEventTarget'
 export * from './emulate/mockRequestIdleCallback'
 export * from './emulate/mockTelemetry'
